First working version of Geodesics
This uses Vincenty's forward and inverse formulae to compute
great circles on an ellipsoid of rotation.  By default, the
ellipsoid of WGS84 is used, but spherical calculations can be
performed if flattening is set to 0.

Additional convenience functions are provided if only some of
the output from the forward and inverse calculations are needed.

The forward and inverse functions are taken from GreatCircle.jl
    https://github.com/acrosby/GreatCircle.jl
by Alexander Crosby, which itself is a port of the pygc package
    https://github.com/axiom-data-science/pygc
by Kyle Wilcox.  These authors are acknowledged in the LICENSE.md
file.